Add UserRepository class
Add the UserRepository class to abstract the data storage.
This class has a ArrayList as attribute to handle the list of users.
By having this repository class, we gain some flexibility. We can later, for
example, change the storage logic (Connect to an external database) without
having to alter other code that the repository classes.